feat(umt_rust): port Iterator utilities from main
- Add `iterator` module with `lazy_map`, `lazy_filter`, `lazy_take`.
- Implement using idiomatic Rust iterators (`enumerate`, `map`, `filter_map`, `take`).
- Add tests in `package/umt_rust/tests/iterator` mirroring `package/main`.
- Fix build error in `tests/array/test_ultra_number_sort.rs` by importing `rand::RngExt`.
