+++ package/umt_rust/tests/iterator/mod.rs
@@ -1,0 +1,3 @@
+pub mod test_lazy_filter;
+pub mod test_lazy_map;
+pub mod test_lazy_take;
